requirements.txt
# 核心依赖
requests>=2.28.0

# douyin-downloader 核心模块依赖
aiohttp>=3.9.0
pyyaml>=6.0.1
gmssl>=3.2.2

# 其他
tqdm>=4.65.0